Move InputBar to Items directory
