+++ client/src/app/styles/index.scss
@@ -27,7 +27,7 @@
 @import '../../common/components/MessageBox.scss';
 @import '../../common/components/Preloader.scss';
 @import '../../modules/list/components/ItemsContainer';
-@import '../../modules/list/components/InputBar/InputBar.scss';
+@import '../../modules/list/components/Items/InputBar';
 @import '../../common/components/Toolbar/components/UserBar/UserBar.scss';
 @import '../../common/components/DropdownMenu.scss';
 @import '../../common/components/SortBox.scss';
