+++ client/src/modules/list/components/Items/InputBar/InputBar.scss
